+++ beehive/apps/project_manager/apis.py
@@ -1,14 +1,23 @@
 from rest_framework.viewsets import GenericViewSet
 from rest_framework import mixins
 
-from apps.project_manager.models import Project, Sprint, Task, ProjectMember
+from apps.project_manager.models import (
+    Project,
+    Sprint,
+    Task,
+    ProjectMember,
+    ProjectRole,
+    ProjectPermission,
+)
 from apps.project_manager.serializers import (
     ProjectSerializer,
     SprintSerializer,
     TaskSerializer,
     ProjectMemberSerializer,
+    ProjectRoleSerializer,
+    ProjectPermissionSerializer,
 )
-from apps.base.mixins import MultiSerializerMixin
+from apps.project_manager.perissions import ProjectMemberPermission
 
 
 class ProjectViewSet(
@@ -21,11 +30,49 @@ class ProjectViewSet(
     queryset = Project.objects.all()
     serializer_class = ProjectSerializer
 
+    def get_permissions(self):
+        permissions = super().get_permissions()
+        if self.action in ["update", "partial_update"]:
+            permissions += [ProjectMemberPermission()]
+        return permissions
+
     def get_queryset(self):
         queryset = super().get_queryset()
         return queryset.filter(
             member__in=[self.request.user],
         )
+
+
+class ProjectPermissionViewSet(mixins.ListModelMixin, GenericViewSet):
+    queryset = ProjectPermission.objects.all()
+    serializer_class = ProjectPermissionSerializer
+
+
+class ProjectRoleViewSet(
+    mixins.ListModelMixin,
+    mixins.RetrieveModelMixin,
+    mixins.CreateModelMixin,
+    mixins.UpdateModelMixin,
+    GenericViewSet,
+):
+    queryset = ProjectRole.objects.all()
+    serializer_class = ProjectRoleSerializer
+
+    def get_permissions(self):
+        permissions = super().get_permissions()
+        permissions += [ProjectMemberPermission()]
+        return permissions
+
+    def get_queryset(self):
+        queryset = super().get_queryset()
+        # Filter project related data that is associated to user
+        return queryset.filter(
+            project=self.kwargs.get("project_id"),
+            project__member__in=[self.request.user],
+        )
+
+    def perform_create(self, serializer):
+        serializer.save(project_id=self.kwargs.get("project_id"))
 
 
 class ProjectMemberViewSet(
@@ -38,8 +85,15 @@ class ProjectMemberViewSet(
     queryset = ProjectMember.objects.all()
     serializer_class = ProjectMemberSerializer
 
+    def get_permissions(self):
+        permissions = super().get_permissions()
+        if self.action in ["create", "update", "partial_update"]:
+            permissions += [ProjectMemberPermission()]
+        return permissions
+
     def get_queryset(self):
         queryset = super().get_queryset()
+        # Filter project related data that is associated to user
         return queryset.filter(
             project=self.kwargs.get("project_id"),
             project__member__in=[self.request.user],
@@ -60,17 +114,22 @@ class SprintViewSet(
     queryset = Sprint.objects.all()
     serializer_class = SprintSerializer
 
+    def get_permissions(self):
+        permissions = super().get_permissions()
+        permissions += [ProjectMemberPermission()]
+        return permissions
+
     filterset_fields = ("project",)
 
     def get_queryset(self):
         queryset = super().get_queryset()
+        # Filter member related data
         return queryset.filter(
             project__member__in=[self.request.user],
         )
 
 
 class TaskViewSet(
-    MultiSerializerMixin,
     mixins.ListModelMixin,
     mixins.RetrieveModelMixin,
     mixins.CreateModelMixin,
@@ -80,10 +139,16 @@ class TaskViewSet(
     queryset = Task.objects.all()
     serializer_class = TaskSerializer
 
+    def get_permissions(self):
+        permissions = super().get_permissions()
+        permissions += [ProjectMemberPermission()]
+        return permissions
+
     filterset_fields = ("sprint",)
 
     def get_queryset(self):
         queryset = super().get_queryset()
+        # Filter member related data
         return queryset.filter(
             sprint__project__member__in=[self.request.user],
         )