+++ beehive/apps/project_manager/urls.py
@@ -8,6 +8,8 @@ from apps.project_manager.apis import (
     SprintViewSet,
     TaskViewSet,
     ProjectMemberViewSet,
+    ProjectRoleViewSet,
+    ProjectPermissionViewSet,
 )
 
 
@@ -18,8 +20,17 @@ default_router.register(
     ProjectMemberViewSet,
     basename="project_members",
 )
+default_router.register(
+    r"project/(?P<project_id>[^/.]+)/roles",
+    ProjectRoleViewSet,
+    basename="project_role",
+)
 default_router.register("project", ProjectViewSet, basename="project")
 default_router.register("sprint", SprintViewSet, basename="sprint")
 default_router.register("task", TaskViewSet, basename="task")
+default_router.register(
+    "permissions", ProjectPermissionViewSet, basename="project_permission"
+)
+
 
 urlpatterns = default_router.urls